server쪽을 수정했습니다.

+++ src/server/index.tsx
@@ -1,3 +1,5 @@
+import dotenv from 'dotenv'
+dotenv.config({ path: '/.env' })
 import * as fs from 'fs'
 import Koa from 'koa'
 import Router from 'koa-router'
@@ -7,7 +9,10 @@ import { createPool, sql } from 'slonik'
 import bcrypt from 'bcryptjs'
 import jwt, { SignOptions } from 'jsonwebtoken'
 
+// const SECRET_KEY = process.env.SECRET_KEY
+
 const SECRET_KEY = 'MySecretKey1$1$234'
+const url = 'postgresql://todo_user:todo_user@localhost/user_inform'
 
 if (!SECRET_KEY) {
   const error = new Error('InvalidSecretKeyError')
@@ -17,7 +22,7 @@ if (!SECRET_KEY) {
 
 const generateToken = (payload: any, options: SignOptions): Promise<string> => {
   const jwtOptions: SignOptions = {
-    issuer: 'songc.io', // 출처명
+    issuer: 'mins', // 출처명
     expiresIn: '7d', // 유효기간
     ...options,
   }
@@ -41,11 +46,14 @@ app.use(bodyParser())
 
 const router = new Router()
 
-const pool = createPool(
-  'postgresql://todo_user:todo_user@localhost/user_inform'
-)
+//const url = process.env.DB_URL
+if (!url) {
+  throw console.error('DB URL이 잘못되었습니다.')
+}
+const pool = createPool(url)
+
 pool.connect(async (connection) => {
-  const data = await connection.query(sql`SELECT * FROM users`)
+  await connection.query(sql`SELECT * FROM users`)
 })
 
 router.get('/', async (ctx) => {
@@ -69,13 +77,15 @@ router.post('/login', async (ctx) => {
       sql`select user_id, password from users where email=${email}`
     )
   ).rows[0]
-  const check = await bcrypt.compare(password, savedPasswordAndUserId.password)
+
+  const check = await bcrypt.compare(
+    password,
+    String(savedPasswordAndUserId?.password)
+  )
   if (check) {
-    // accessToken을 사용했더니 안보임..?
     const accessToken = await generateToken(
-      { user_id: savedPasswordAndUserId.user_id },
-      // review: 지금은 1시간만에 로그인이 풀립니다.
-      { subject: 'access_token', expiresIn: '1h' }
+      { user_id: savedPasswordAndUserId?.user_id },
+      { subject: 'access_token', expiresIn: '1d' }
     )
     ctx.cookies.set('access_token', accessToken, {
       httpOnly: true,
@@ -87,31 +97,30 @@ router.post('/login', async (ctx) => {
   }
 })
 
-// user Id를 전달하기 위해 만듬
-router.get('/user/:loginEmail', async (ctx) => {
-  const { loginEmail } = ctx.params
-  const userId = await pool.query(
-    sql`select id from users where email=${loginEmail}`
-  )
-  ctx.body = JSON.stringify(userId)
-})
+// // user Id를 전달하기 위해 만듬
+// router.get('/user/:loginEmail', async (ctx) => {
+//   const { loginEmail } = ctx.params
+//   const userId = await pool.query(
+//     sql`select id from users where email=${loginEmail}`
+//   )
+//   ctx.body = JSON.stringify(userId)
+// })
 
-router.get('/chatting', async (ctx) => {
-  // review: 코드상에선 chatting인데 테이블 이름은 text로 되어 있는 것 같습니다. 가능하면 일치시켜 주세요.
+router.get('/text', async (ctx) => {
   const data = await pool.query(sql`SELECT * FROM text ORDER BY date DESC`)
   ctx.response.body = JSON.stringify(data)
 })
 
 // review: 유저 아이디를 직접 받는 것은 잘못된 방식입니다. ctx의 cookie로부터 유저 정보를 알아내야 합니다.
-router.post('/chatting/:user_id', async (ctx) => {
-  const { user_id } = ctx.params
+router.post('/text', async (ctx) => {
   const { text } = ctx.request.body
-  pool.query(sql`insert into text(user_id,text) values (${user_id},${text})`)
+  pool.query(sql`insert into text(text) values (${text})`)
   ctx.status = 200
 })
 
-router.delete('/chatting/:id', async (ctx) => {
+router.delete('/text/:id', async (ctx) => {
   const { id } = ctx.params
+  if (id === undefined) return
   pool.query(sql`delete from text where id=${id}`)
   ctx.status = 200
 })
